Conflicts:
	engine-spark/src/main/scala/com/intel/intelanalytics/engine/spark/SparkOps.scala
	engine-spark/src/test/scala/com/intel/intelanalytics/engine/spark/frame/plugins/SparkJoinITest.scala

+++ interfaces/src/main/scala/com/intel/intelanalytics/algorithm/QuantileTarget.scala
@@ -24,8 +24,8 @@
 package com.intel.intelanalytics.algorithm
 
 /**
- * quantile target that will be applied to an element
+ * Quantile target that will be applied to an element
  * @param quantile quantile. For eg, 40 means 40th quantile
  * @param weight weight that will be applied to the element
  */
-case class QuantileTarget(quantile: Int, weight: BigDecimal)
+case class QuantileTarget(quantile: Double, weight: BigDecimal)
